fix: losslessly validate numeric indices — method params + type/heap fields (index-hardening H1)
napi decodes a `u32` param/field via ECMAScript ToUint32 BEFORE any Rust range
check (2**32->0, -1->u32::MAX, NaN/Infinity->0, a fraction truncates), so an
out-of-domain numeric INDEX silently aliases a valid index and emits
semantically-wrong wasm. Add a shared `checked_index(n: f64, what) -> Result<u32>`
that reads the exact IEEE double (no coercion) and rejects a non-finite /
non-integer / negative / >u32::MAX value catchably.

PART A: 16 method params `u32`->`f64` / `Vec<u32>`->`Vec<f64>`, validated at the
method top — the 11 collection `getByIndex`, plus `buildFunction.argLocalIndices`
and `replaceExportedFunc`/`replaceImportedFunc` `funcIndex`+`argLocalIndices`. The
arena resolvers still take `u32`, fed the validated output.

PART B: 6 enum/object index fields `u32`->`f64` (HeapType Concrete/Exact
`typeIndex`, HeapType RecGroup `recIndex`, RecGroupRef RecGroup `recIndex`,
RecGroupRef Existing `typeIndex`, BlockType MultiValue `typeIndex`), validated at
every WRITE consume site before the range-check/`resolve_type_id`; the READ paths
set from walrus flip `as u32`->`as f64`. Closes the B5c review finding.

Both `u32` and `f64` render as TS `number`, so `index.d.ts` is byte-identical.
InstrDesc/CatchClause scalar index fields are untouched (H2). Adds 10 negative
tests; native + WASI suites 368/368.

+++ src/valtype.rs
@@ -60,15 +60,15 @@ pub enum HeapType {
   Abstract { kind: AbstractHeapType },
   /// A concrete (indexed) heap type: `(ref $t)`, referencing a defined type by
   /// its stable index.
-  Concrete { type_index: u32 },
+  Concrete { type_index: f64 },
   /// An exact heap type: `(ref exact $t)` (custom-descriptors proposal),
   /// referencing a defined type by its stable index.
-  Exact { type_index: u32 },
+  Exact { type_index: f64 },
   /// An intra-rec-group reference: `(ref $sibling)` to the `rec_index`-th
   /// member of the rec group currently being built. Valid ONLY inside an
   /// `addRecGroup` member descriptor (a sibling has no stable arena index yet);
   /// every other converter rejects it with a catchable error.
-  RecGroup { rec_index: u32 },
+  RecGroup { rec_index: f64 },
 }
 
 /// An abstract heap type, mirroring `walrus::AbstractHeapType` 1:1.
@@ -179,12 +179,12 @@ pub enum RecGroupRef {
   /// A sibling member of the rec group currently being built, by position.
   RecGroup {
     /// The position of the referenced member in the `members` array.
-    rec_index: u32,
+    rec_index: f64,
   },
   /// An existing type in the module, by its stable arena index.
   Existing {
     /// The stable `.index()` of the referenced existing type.
-    type_index: u32,
+    type_index: f64,
   },
 }
